requirements.txt
openai
langchain
langchain-openai
streamlit
python-dotenv